requirements.txt
streamlit>=1.32.0
ollama>=0.4.0
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
python-docx>=1.1.0
Pillow>=10.0.0
pytesseract>=0.3.10
fpdf2>=2.8.0
pandas>=2.2.0
